third_party/blink/web_tests/external/wpt/custom-elements/registries/ShadowRoot-init-customElementRegistry-expected.txt

This is a testharness.js-based test.
[FAIL] A newly attached disconnected ShadowRoot should use the global registry by default
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] A newly attached connected ShadowRoot should use the global registry by default
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] A newly attached disconnected ShadowRoot should use the scoped registry if explicitly specified in attachShadow
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] A newly attached connected ShadowRoot should use the scoped registry if explicitly specified in attachShadow
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] attachShadow() should throw for a null customElementRegistry value
  assert_throws_js: function "() => host.attachShadow({mode: 'closed', customElementRegistry: null})" did not throw
Harness: the test ran to completion.
